REFACT create services for the generator pseudo
